+++ src/main/java/com/example/filevault/util/UserWorkUtils.java
@@ -1,8 +1,5 @@
 package com.example.filevault.util;
 
-import com.example.filevault.config.security.UserRole;
-import com.example.filevault.dao.UserDao;
-import com.example.filevault.entity.UserEntity;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
